move client validation to platform
