+++ internal/platform/terminal/terminal_client.go
@@ -2,8 +2,11 @@ package terminal
 
 import (
 	"errors"
+	"fmt"
 	"net/http"
+	"strings"
 
+	"github.com/scaleway/scaleway-cli/v2/internal/platform"
 	"github.com/scaleway/scaleway-sdk-go/logger"
 	"github.com/scaleway/scaleway-sdk-go/scw"
 	"github.com/scaleway/scaleway-sdk-go/validation"
@@ -78,10 +81,114 @@ func (p *Platform) CreateClient(httpClient *http.Client, configPath string, prof
 		return nil, err
 	}
 
-	return client, nil
+	return client, validateClient(client)
 }
 
 func errIsConfigFileNotFound(err error) bool {
 	var target *scw.ConfigFileNotFoundError
 	return errors.As(err, &target)
+}
+
+// configErrorDetails generate a detailed error message for an invalid client option.
+func configErrorDetails(configKey, varEnv string) string {
+	// TODO: update the more info link
+	return fmt.Sprintf(`%s can be initialised using the command "scw init".
+
+After initialisation, there are three ways to provide %s:
+- with the Scaleway config file, in the %s key: %s;
+- with the %s environement variable;
+
+Note that the last method has the highest priority.
+
+More info: https://github.com/scaleway/scaleway-sdk-go/tree/master/scw#scaleway-config`,
+		configKey,
+		configKey,
+		configKey,
+		scw.GetConfigPath(),
+		varEnv,
+	)
+}
+
+// validateClient validate a client configuration and make sure all mandatory setting are present.
+// This function is only call for commands that require a valid client.
+func validateClient(client *scw.Client) error {
+	accessKey, _ := client.GetAccessKey()
+	if accessKey == "" {
+		return &platform.ClientError{
+			Err:     fmt.Errorf("access key is required"),
+			Details: configErrorDetails("access_key", "SCW_ACCESS_KEY"),
+		}
+	}
+
+	if !validation.IsAccessKey(accessKey) {
+		return &platform.ClientError{
+			Err: fmt.Errorf("invalid access key format '%s', expected SCWXXXXXXXXXXXXXXXXX format", accessKey),
+		}
+	}
+
+	secretKey, _ := client.GetSecretKey()
+	if secretKey == "" {
+		return &platform.ClientError{
+			Err:     fmt.Errorf("secret key is required"),
+			Details: configErrorDetails("secret_key", "SCW_SECRET_KEY"),
+		}
+	}
+
+	if !validation.IsSecretKey(secretKey) {
+		return &platform.ClientError{
+			Err: fmt.Errorf("invalid secret key format '%s', expected a UUID: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx", secretKey),
+		}
+	}
+
+	defaultOrganizationID, _ := client.GetDefaultOrganizationID()
+	if defaultOrganizationID == "" {
+		return &platform.ClientError{
+			Err:     fmt.Errorf("organization ID is required"),
+			Details: configErrorDetails("default_organization_id", "SCW_DEFAULT_ORGANIZATION_ID"),
+		}
+	}
+
+	if !validation.IsOrganizationID(defaultOrganizationID) {
+		return &platform.ClientError{
+			Err: fmt.Errorf("invalid organization ID format '%s', expected a UUID: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx", defaultOrganizationID),
+		}
+	}
+
+	defaultZone, _ := client.GetDefaultZone()
+	if defaultZone == "" {
+		return &platform.ClientError{
+			Err:     fmt.Errorf("default zone is required"),
+			Details: configErrorDetails("default_zone", "SCW_DEFAULT_ZONE"),
+		}
+	}
+
+	if !validation.IsZone(defaultZone.String()) {
+		zones := []string(nil)
+		for _, z := range scw.AllZones {
+			zones = append(zones, string(z))
+		}
+		return &platform.ClientError{
+			Err: fmt.Errorf("invalid default zone format '%s', available zones are: %s", defaultZone, strings.Join(zones, ", ")),
+		}
+	}
+
+	defaultRegion, _ := client.GetDefaultRegion()
+	if defaultRegion == "" {
+		return &platform.ClientError{
+			Err:     fmt.Errorf("default region is required"),
+			Details: configErrorDetails("default_region", "SCW_DEFAULT_REGION"),
+		}
+	}
+
+	if !validation.IsRegion(defaultRegion.String()) {
+		regions := []string(nil)
+		for _, z := range scw.AllRegions {
+			regions = append(regions, string(z))
+		}
+		return &platform.ClientError{
+			Err: fmt.Errorf("invalid default region format '%s', available regions are: %s", defaultRegion, strings.Join(regions, ", ")),
+		}
+	}
+
+	return nil
 }
